style: Renamed files for better naming conventions
